% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcofoG.R
\name{article}
\alias{article}
\title{Article}
\usage{
article(..., md_extensions = c("-autolink_bare_uris"))
}
\description{
Formatage d'un article pour l'auto-archivage
}
\details{
La fonction est appelée par le modèle Markdown article
}
