% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Automap.R
\name{Automap}
\alias{Automap}
\title{AutoMap}
\usage{
Automap()
}
\description{
Interface Shiny de création automatique de cartes de terrain
}
\details{
Basée sur la base de données Guyafor
}
\author{
Elie Guedj, \email{elie.guedj@ecofog.gf}
}
