% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guyafor.R
\name{Guyafor2df}
\alias{Guyafor2df}
\title{Donnees Guyafor}
\usage{
Guyafor2df()
}
\description{
Importation de l'ensemble des données de la base Guyafor dans un dataframe
}
\details{
La fonction exécute une requête sur le serveur sql.ecofog.gf pour lire les données
}
\author{
Gaëlle Jaouen, \email{gaelle.jaouen@ecofog.gf}
}
