% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guyafor.R
\name{Paracou2df}
\alias{Paracou2df}
\title{Donnees Paracou}
\usage{
Paracou2df()
}
\description{
Importation de l'ensemble des données de Paracou de la base Guyafor dans un dataframe
}
\details{
La fonction exécute une requête sur le serveur sql.ecofog.gf pour lire les données
}
