% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guyafor.R
\docType{data}
\name{Paracou6_2016}
\alias{Paracou6_2016}
\title{Inventaire 2016 de la parcelle 6 de Paracou.}
\format{
Un tibble contenant une mesure (année 2016) d'un arbre par ligne
}
\source{
<http://paracou.cirad.fr/>
}
\usage{
Paracou6_2016
}
\description{
Jeu de données public extrait de la base Guyafor.
}
\examples{
# Nombre d'arbres
nrow(Paracou6_2016)
}
\keyword{datasets}
