% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoFoG.R
\name{Tricoter}
\alias{Tricoter}
\alias{TricoterTout}
\alias{TricoterArticle}
\alias{TricoterPresentation}
\alias{TricoterOuvrage}
\alias{TricoterMemo}
\title{Tricoter}
\usage{
TricoterTout(destination = "docs")

TricoterArticle(destination = "docs")

TricoterPresentation(destination = "docs")

TricoterOuvrage(destination = "docs")

TricoterMemo(destination = "docs")
}
\arguments{
\item{destination}{Dossier de destination des documents}
}
\description{
Création des tous les documents à partir des modèles
}
\details{
Utilisé pour vérifier le bon fonctionnement des modèles
}
