% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoFoG.R
\name{TricoterTout}
\alias{TricoterTout}
\title{TricoterTout}
\usage{
TricoterTout(destination = "docs")
}
\arguments{
\item{destination}{Dossier de destination des documents}
}
\description{
Création des tous les documents à partir des modèles
}
\details{
Utilisé pour vérifier le bon fonctionnement des modèles
}
