% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoFoG.R
\name{article}
\alias{article}
\title{Article}
\usage{
article(..., md_extensions = c("-autolink_bare_uris"))
}
\arguments{
\item{...}{Arguments optionnels passés à \code{\link{pdf_document}}}

\item{md_extensions}{Extensions markdodown, cf. \code{\link{pdf_document}}}
}
\description{
Formatage d'un article pour l'auto-archivage
}
\details{
La fonction est appelée par le modèle Markdown article
}
