% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Automap.R
\name{Automap}
\alias{Automap}
\title{AutoMap}
\usage{
Automap(
  data = NULL,
  UID = NULL,
  PWD = NULL,
  Driver = "SQL Server Native Client 10.0"
)
}
\arguments{
\item{data}{Dataframe in Guyafor format}

\item{UID}{Compte SQL Server utilisé pour la connexion. Par défaut, l'authentification est faite par Windows.}

\item{PWD}{Mot de passe du compte SQL Server.}

\item{Driver}{Nom du pilote ODBC, dépendant de l'installation du système. La valeur par défaut est le nom le plus fréquent. "SQL Server" peut fonctionner si la valeur par défaut renvoie une erreur indiquant que le pilote ODBC n'existe pas.}
}
\description{
Interface Shiny de creation automatique de cartes de terrain
}
\details{
Basee sur la base de donnees Guyafor
}
\author{
Elie Guedj, \email{elie.guedj@ecofog.gf}
}
