% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Automap.R
\name{Automap}
\alias{Automap}
\title{AutoMap}
\usage{
Automap(data = NULL, UID = NULL, PWD = NULL)
}
\arguments{
\item{data}{Dataframe in Guyafor format}

\item{UID}{Username for SQL Server connection}

\item{PWD}{Password for SQL Server connection}
}
\description{
Interface Shiny de creation automatique de cartes de terrain
}
\details{
Basee sur la base de donnees Guyafor
}
\author{
Elie Guedj, \email{elie.guedj@ecofog.gf}
}
