% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Guyafor.R
\name{Guyafor2df}
\alias{Guyafor2df}
\title{Donnees Guyafor}
\usage{
Guyafor2df(WHERE = NULL)
}
\arguments{
\item{WHERE}{Clause WHERE optionnelle de la requête SQL envoyée au serveur}
}
\value{
Un dataframe contenant le résultat de la requête ODBC
}
\description{
Importation de l'ensemble des données de la base Guyafor dans un dataframe
}
\details{
La fonction exécute une requête sur le serveur sql.ecofog.gf pour lire les données.
Un pilote ODBC pour SQL Server doit être accessible.
}
\examples{
if (!any(is.na(pingr::ping_port("sql.ecofog.gf", port=1433))))
  # Si le serveur sql.ecofog.gf est accessible
  Paracou15 <- Guyafor2df(WHERE="Forest='Paracou' AND Plot='15' AND CensusYear=2016")
}
\author{
Gaëlle Jaouen, \email{gaelle.jaouen@ecofog.gf}
}
